refactor(backend/utils): Move password hashing functions to utils and update imports
- Moved `hashPassword` and `verifyPassword` functions from `passwordHashService` to `passwordHashUtils`.
- Updated all imports in `userService.js` and `userControllers.js` to use `passwordHashUtils` for password hashing and verification.
- Ensured all password-related operations are handled by the utility functions.
